Changed from using the Paper API to Spigot API because SpigotMC doesn't accept Paperspigot based plugins :( but its ok because it all works fine with Spigot :)
Added a debug mode which is activated by command to give info about the plugin if its broken somewhere. Many messages are to be added as the bugs come.

Removed the kick players on reload and replaced with just resending the updated pack to all players. This is reload safe (so far) and people dont have to reconnect :)

Removed some kt files that were not being used, mainly extensions from other projects that I added for potential use.

Updated to version 0.7.0